TL-BiRNN-Pick/requirements.txt
altgraph==0.17.4
appdirs==1.4.4
apptools==5.2.0
asn1crypto==0.24.0
backcall==0.2.0
CacheControl==0.12.14
cachy==0.3.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cleo==0.8.1
clikit==0.6.2
colorama==0.4.5
conda==4.5.1
configobj==5.0.8
crashtest==0.3.1
cryptography==2.2.2
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
distlib==0.4.0
docopt==0.6.2
einops==0.4.1
entrypoints==0.4
envisage==6.1.1
filelock==3.4.1
googledrivedownloader==0.4
html5lib==1.1
idna==2.6
imageio==2.16.0
importlib-metadata==1.7.0
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
isodate==0.6.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.1
keyring==23.4.1
kiwisolver==1.3.1
lockfile==0.12.2
MarkupSafe==2.0.1
matplotlib==3.3.4
menuinst==1.4.11
metakernel==0.28.2
msgpack==1.0.5
nest-asyncio==1.6.0
networkx==2.5.1
numpy==1.19.5
oct2py==5.3.0
octave-kernel==0.32.0
packaging==20.9
pandas==1.1.5
parso==0.7.1
pastel==0.2.1
pefile==2024.8.26
pexpect==4.9.0
pickleshare==0.7.5
Pillow==8.4.0
pipreqs==0.4.11
pkginfo==1.10.0
platformdirs==2.4.0
plotly==5.14.1
poetry==1.1.15
poetry-core==1.0.8
progressbar==2.5
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pyarrow==1.0.0
pycosat==0.6.3
pycparser==2.18
pyekfmm==0.0.8.7
pyface==7.4.4
Pygments==2.14.0
pyinstaller-hooks-contrib==2022.0
pylev==1.4.0
pyOpenSSL==17.5.0
pyparsing==3.0.7
PySocks==1.6.8
python-dateutil==2.8.2
pytz==2024.1
pyvista==0.33.3
pywin32==223
pywin32-ctypes==0.2.3
PyYAML==6.0.1
pyzmq==25.1.2
rdflib==5.0.0
requests==2.18.4
requests-toolbelt==0.9.1
ripleyk==0.0.3
ruamel-yaml==0.15.35
scikit-learn==0.24.2
scipy==1.5.4
scooby==0.5.11
shellingham==1.5.0
six==1.11.0
tenacity==8.2.2
threadpoolctl==3.1.0
tomlkit==0.11.6
torch-geometric==2.0.3
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
traits==6.4.1
traitsui==7.4.3
typing-extensions==4.1.1
urllib3==1.22
virtualenv==20.17.1
vtk==9.2.6
wcwidth==0.2.13
webencodings==0.5.1
win-inet-pton==1.0.1
wincertstore==0.2
yacs==0.1.8
yarg==0.1.10
zipp==3.6.0
